refactor: move UpdateInfo to actioncard
