+++ packages/frontend/src/components/common/ActionCard.tsx
@@ -1,6 +1,8 @@
-import { UpdatedByUser } from 'common';
-import React, { Dispatch, SetStateAction } from 'react';
-import { UpdatedInfo } from '../Home/LatestSavedCharts';
+import { Colors } from '@blueprintjs/core';
+import { SessionUser, UpdatedByUser } from 'common';
+import React, { Dispatch, FC, SetStateAction } from 'react';
+import styled from 'styled-components';
+import { useTimeAgo } from '../../hooks/useTimeAgo';
 import LinkButton from './LinkButton';
 import ModalActionButtons from './modal/ModalActionButtons';
 
@@ -9,6 +11,37 @@ type ActionCardProps<T> = {
     url: string;
     setActionState: Dispatch<SetStateAction<{ actionType: number; data?: T }>>;
     isChart?: boolean;
+};
+
+export const UpdatedLabel = styled.p`
+    color: ${Colors.GRAY2};
+    font-size: 0.75em;
+    font-weight: 400;
+    margin-top: 0.38em;
+    line-height: 12px;
+`;
+
+export const UpdatedInfo: FC<{
+    updatedAt: Date;
+    user: Partial<SessionUser> | undefined;
+}> = ({ updatedAt, user }) => {
+    const timeAgo = useTimeAgo(updatedAt);
+
+    return (
+        <UpdatedLabel>
+            Last edited <b>{timeAgo}</b>{' '}
+            {user && user.firstName ? (
+                <>
+                    by{' '}
+                    <b>
+                        {user.firstName} {user.lastName}
+                    </b>
+                </>
+            ) : (
+                ''
+            )}
+        </UpdatedLabel>
+    );
 };
 
 const ActionCard = <
